feat: implement promo code discount and update cart item UI

+++ feature-cart/data/src/main/java/br/mrenann/cart/data/mapper/CartMapper.kt
@@ -44,7 +44,7 @@ fun List<CartItemEntity>.toProductCart() = map { entity ->
     ProductCart(
         id = entity.productId.toInt(),
         title = entity.name,
-        price = entity.price.toInt(),
+        price = entity.price,
         description = "",
         category = Category(
             id = 1,
